Fix incorrect comparision between symbols and strings in deadlines
In register_deadline function, we have a logic where we don't allow extending
the deadline if the target is same. However, we were comparing symbols with
strings to decide if the target is same, which always results in us deciding
the targets are different and extending the deadline. This commit fixes This
by using strings for deadline targets.