Support for Postgres in burstable VMs
Added configuration to allow Postgres server to be hosted inside Burstable instances. The core of this change is in option.rb file, where new PG configurations are added. One code change was needed to set a minimum value of work_mem to 1MB for the smaller instances. Also added billing support for PG, similar to what was done earlier for VMs. New billing resource_types based on CPU percent are added and used when the corresponding VM uses VmCpuPercent counter. I refactored the billing code for VM a bit to make it easier to load the right billing rate for Postgres.

Fixed a typo in biling rates for leaseweb-* locations for PostgresCores.

Updated tests.