Expand `r` in `util.rb` to get a list of expected exit statuses
By default, `r` threw exceptions whenever the exit status was not 0.
However, in some scenarios the caller is ok with some exit statuses. For
example, when the caller wants to stop a systemd unit if it exists. In
which case, it is ok if the exit status is 5.

This change adds the `expect:` parameter to `r` to allow for that:

```
def r(commandline, stdin: "", expect: [0])
```