Make page.refresh work in specs
The default Capybara page.refresh submits the rack request for the
previous environment, which does not work with Rack::Lint.  It also
causes issues with cookies.  When you refresh a real browser, it
will submit using the current cookie jar, it will not blindly use
the same Cookie header as the previous request.

Use page.refresh in a couple places we were purposely avoiding it
previously.