Support Inference{Endpoint,Token} as ObjectTag members
This makes the access control system correctly support permissions on
these objects, which were added after work on the new access system
began.

Because InferenceToken doesn't exist as a separate class, and the model
class (ApiKey) uses the TYPE_ETC ubid type, this is more complicated
than it otherwise would be.  I added a hack to UBID.resolve_map to treat
ubids starting with "et" as ApiKey.  If we every want to use another
model class that uses the TYPE_ETC ubid type, this approach will not
scale well (we can hack it to work, but you then need to query on all
possible tables it could be).

For consistency with other models which define name, add ApiKey#name and
have it return the ubid.