Fix updating host used memory
When creatign a burstable VM, the incorrect amount of memory was passed to the VmHostAllocation, resulting in adding only memory used by the VM to used_hugepages_1g on the host. Instead, we should use the memory amount allocated to the slice. When slice is destroyed it updates the host to free the memory it allocated.

The host is updated for the whole slice even though not all that memory is immediately used by the VMs. That is because that memory is now reserved for that slice and cannot be used by other allocations. Furture allocations that place the memory in the VM will update the Slice memory counters until the slice is full, and the host memory counters will not change.

The mem_gib_ratio method was only used once in the Allocator so I moved it to be a custom method of a Request structure. Moved the tests accordingly.

New expect() conditions were added in the existing tests to detect this bug.