Migration for memory_1g to memory_gib rename
As part of the naming convention cleanup for memory-related variables we are renaming memory_1g columns on VmHostSlice to memory_gib. This commit is the schema migration part split into two parts - one introducing the columns, the other moving the values and constraints to the new ones.
Code changes will follow.