Cap the cleanup cache entries at 200 for each run
While we clean up old cache entries, we also remove them from blob
storage, which can take some time.

We have a 2-minute limit in respirate for each run. If it exceeds this
limit, respirate considers the run stuck and terminates itself.

We encountered this issue in production when we needed to clean up over
1000 cache entries in one run, which took more than 2 minutes. As a
result, respirate crashed multiple times.

Since we run the cleanup job every 5 minutes, we can cap the cleanup
cache entries at 200 for each run. This way, we can avoid the respirate
crash and also keep the cache clean.