Remove monthly cap
We used to cap the number of days in month at 28 for billing purposes. This
means your bill would increase linearly till 28th day but then stay flat till
the end of the month. It is useful to keep the total price same every month
independent from the number of days in the month. This special handling
complicates our billing calculations, so we decided to remove it and now the
amount we would charge increase linearly till the end of the month. We reduced
our rates (in previous few commits) to compensate for this change.