Support unrestricting access for tokens
With the new token access control, it's not obvious how to unrestrict
a token after restricting it.  Previously, you could add an entry for
all actions and all objects, but such entries are now ignored, since
they are no different from the empty case, and we don't want to
automatically create such an entry in the empty case (that would be
a fail-open design).

A restricted token that is allowed all actions on all objects is the
same as an unrestricted token, so this allows users who want to
unrestrict their tokens an easy way to do so.