Return error json without error key for runtime endpoints
Returning an error json without an error key on top for runtime endpoints.
It is required because docker layer caching expects error in this format to
skip duplicate cache entries. Passing it in this way works properly
with the cache proxy as we directly pass the error to the client
without parsing it.