Relocate the tag creation form to the top of the list
Having lots of pages with limited functionality can complicate the user
experience. We could merge some of these pages. Placing single input
forms at the top of the listing page is a common web pattern. It is more
user-friendly as it allows users to create a new tag without the need to
navigate between pages.

I've also repositioned the 'Membership' button from the left to the
right, placing it next to the 'Remove' button. I believe grouping all
action buttons in one column enhances usability. I've renamed it to
'Manage' and will be improving this management page in upcoming commits.