Use a template database for much faster migrating of test databases used for parallel testing
Previously, each parallel database was migrated separately from
scratch.  This switches to only migrating the clover_test database,
and using the createdb -T clover_test option to clone that to
create the parallel test databases.  This is much faster, and
doesn't clutter the screen with tons of unnecessary output,
for easier debugging when testing migrations.

With this change, the migrate proc no longer needs a db keyword,
so remove that and use DB instead.