Consolidate tag detail operations into one page
I previously moved the create tag form to the top of the tag listing.

We can also move the update tag form to the top of the detail page. In
other words, I moved membership operations to the details page.

I didn't realize the tag name is clickable. I might change its color,
but the "Manage" button already has this link now.