Show all subject tags in the policy selector
If the current user doesn't have permission for a subject tag, it won't
appear in the policy selector.

This is an issue because the selector doesn't provide an option for that
tag, which is interpreted as no policy option. We should display all
subject tags in the policy selector, but if the user lacks permission to
add one of them, it should be disabled.