Improve UI for token restriction
This improvement has similar characteristics to the previous one. I
tried to reduce the number of pages.

I also moved the restrict and unrestrict actions to a separate card, so
users can get more focused information about token restriction.

Additionally, I believe we don't need to show information about our
token access policy feature if the token already has entries. If it has
entries, it means the user is already aware of the feature. They can
read more about it in the documentation.