Allow existing users to connect their social accounts
In the previous commit, we prevented existing users from connecting
their social accounts without logging in first for security reasons.

This PR adds a new tab to the "My Account" page that allows existing
users to connect and disconnect their social accounts.

If the email of the current account does not match the social account,
we do not allow the operation.

We also do not allow disconnecting the last login method. However, we
can allow disconnecting the last omniauth provider if the user has a
password. The `has_password?` method is not yet public in Rodauth, so I
created a PR for it.