Do not link social identities to existing accounts without logging in
If an account already exists and it's being accessed for the first time via
social identity, we should not automatically link it to this identity. Existing
account holders should be able to link their social identites only after
they've logged in. This way, we can ensure that users are able to securely
connect their accounts. In the current situation, if I don't link my GitHub
account to Ubicloud and my GitHub account gets hacked, the attacker can log in
to Ubicloud. However, in the proposed situation, the attacker would also need
to know my Ubicloud password. It's not a case for newly created accounts.