Reduce the number of queries to modify subject tags from user page
When updating policies for existing users and invitation:

* Avoid queries inside the loop over the provided parameters.
  Instead, run the following queries before the loops:

  * Get array of account ids in the submission that are also in
    the project.
  * Get a hash of all subject tags in the project that those
    accounts are in.
  * Get a hash of project invitations for the project that match
    submitted emails

  Using the hashes, and using SubjectTag instance instead of tag
  name as the key additions and removals, avoid the following N+1
  queries in loops:

  * Get user
  * Get subject tags for each submitted user
  * Get subject tag for each tag being added to
  * Get subject tag for each tag being removed from
  * Get invitation
  * Update invitation

  For the invitations, this now does one query per policy being
  set, updating all invitations that need to use that policy.

This avoids running the loop code for submitted accounts that are
not associated with the project. I think the previous code would
let a project admin add non-project users to subject tags.  I
guess the project admins could technically already do that because
inviting an existing user is really a force-join and not an
invitation, but hopefully we can fix that later.

When inviting a user, skip the subject tag query if not inviting
them into a subject tag.

To reduce flash error size, combine duplicate issues in the error
message.

This adds a new issue type for the case where the user to be
updated has been removed from the project, and adds a test for
that.