Avoid 2*N+1 query issue when displaying project user page
The policy_select control was issues 2 separate queries
per account, one to get the tags allowed to add, and one
to get all subject tags.  These are the same for every user,
so get them in the route code.  Instead of showing all
subject tags, only show subject tags the user is allowed
to view.