Run specs in verbose mode by default
I noticed that we weren't running specs in verbose mode by default.
This hid a fairly important issue in rodauth-omniauth that I only
discovered when running `rake spec_separate`.  This switches to
running specs in verbose mode by default. This moves the warning
filters before the requiring of other libraries in spec_helper.
It adds a couple more warning filters, to filter out a couple of
verbose-mode only warnings.  I did not filter out the
rodauth-omniauth warning, since that is a real bug that needs to
be fixed upstream.

Note that in some cases, rspec will enable verbose warnings by
default, but it does not appear to happen in all cases, and
turbo_tests does not seem to enable verbose warnings by default.