Add migration to use archived_record for applied_{subject,action,object}_tag
This adds a simple trigger function to insert into the
archived_record table, and triggers on the
applied_{subject,action,object}_tag tables to use it.  This
approach is used instead of the before_destroy approach because
these tables do not have models associated with them.

We could consider this approach for other tables in the future
as a more efficient replacement for the before_destroy approach.