Destroy primary only after failover candidate started to work
Previously, we could have destroyed the primary server before the new standby
started to work towards failover, which means if something happens to the
standby (for example gets deleted due to disable HA request), we would delete
both primary and standby servers.

This commit ensures that we only destroy the primary server after the standby
server has started to work towards failover. In next commit, we will also add
extra protection to not delete the standby server if it is picked for failover.