Add stopped label to VM Nexus
We need to stop a few virtual machines every month due to fraudulent
activity before permanently deleting the data.

Currently, we manually stop the VM on the dataplane, which causes an
unavailability page, and users can't see that it's stopped in the UI.

A proper stop may require releasing CPU and memory resources, but the
initial version only stops the VM process without releasing those
resources.

This will help with operational tasks.