Do not overwrite the source port when load balancing
With the PR https://github.com/ubicloud/ubicloud/pull/2272, we have
introduced source port overwriting when load balancing to a neighbor
node. This has created confusion for connection tracking and simply
broke load balancer. In this commit, I am changing the logic back to not
overwriting, but introducing selective destination port overwriting.

This way, if the connection is forwarded to a neighbor node, we are
still using the load balancer public port and keep the source port same.
To be able to process that on the neighbor side, we are overwriting the
destination port from public to private upon arrival. For packs that are
coming from a neighbor node's private IP to the arrival private IP but
in public port.

This creates a problem for connections to be forwarded to the local VM.
The reason is, DNATing can only be done at the preforward hook which is
executed before applying the firewall rules. Therefore, we cannot leave
the public port, we must overwrite it to private port since the
application is running in the private port. Once we overwrite that, we
have to make sure firewall rules for the public port are still applied
to that connection. We will solve that in the next commit.