Add Raw/Preview tabs to playground response section
Add two tabs to the response section of the playground:

1. Raw: This shows the raw response of the model.
2. Preview: A rendered representation of the model response.

While the response is streaming in, only the "Raw" tab is visible. When
the request has finished, the Preview tab becomes available and users
can switch back and forth between the raw response and the rendered
representation.

During request processing, the size of inference_response_raw is
restricted. If the received content becomes larger than that height,
scrollbars are shown that automatically keep their focus on the
bottom. After request processing has finished, we switch to "Preview"
and remove the length restriction from "Raw". So if users switch back to
"Raw" they see the full content without max height restriction.