Fix Subnet Nexus rekeying synchronisation issue
A recent bug report showed that when multiple VMs are provisioned,
sometimes, the private subnet gets stuck in wait_inbound_setup. The main
reason is that, when we start rekeying nics, we include all of the nics
in wait_setup label. The problem with that is, sometimes the VM is not
allocated just yet. If it's not allocated, its public ipv6 address is
not set. Therefore, for tunnels that target this nic, RekeyNicTunnel
prog errors out in subdivide_network method. To solve this, we add a new
state to NicNexus to wait VM allocation, we jump to the next label only
after the allocation is completed.