Support migrating a single migration down
This is useful if testing a PR with a migration version that is
less than the most recent migration version on the main branch.
To use, provide the migration filename as VERSION when running
test_down/dev_down.

While here, fix the test_down/dev_down descriptions to make it
clear that VERSION is required.

It's possible to allow migrating a single migration up, but I'm not
sure there is a need for it, and it requires additional refactoring.

This switches Sequel::Migrator.apply to
Sequel::TimestampMigrator.apply, since we know that we will be
using TimestampMigrator and not IntegerMigrator.