Add put-certificate functionality to setup-cert-server
We used to write certificate and the key remotely into the host.
However, this was an anti pattern because the cert-server functionality
was mostly left to cert_server_setup script. It also made it difficult
to write the cert and key files in a safe way into the correct paths. In
this commit, we are adding the necessary functionality to the
setup-cert-server script.