Refactor setup and stop_and_remove steps into cert_server_setup
Instead of calling 3 functions independently through the
setup-cert-server, we create wrapper functions in cert_server_setup and
make use of them. This way, the setup-cert-server is easier to digest.

We also add a new parameter RestartSec to the systemd service. The main
reason is that, if the VM is attached to a load balancer way before the
necessary interfaces are setup, the metadata-endpoint was quickly
failing and hit the restart count limit. This way, we slow down the
restarts and let it quickly catch up with VM provisioning.

This commit also introduces the tests for the necessary scripts,
previously, there were no tests at all.