Fix VM destroy for a Load Balancer detached VM
If the customer very quickly detaches the VM from the LB and immediately
starts the destroy sequence, because of the async nature of the
scheduling, we might end up in a state that vm.load_balancer returns
nil. That was causing nil ref for vm.load_balancer.evacuate_vm(vm) call.
We safeguard that by moving the code block into an if check.