Map runner label aliases to the actual labels automatically
We have a several labels that serve as an alias for another label.
Currently, we keep the specifications of each label separately, which
causes a lot of duplicate data in the YAML file and makes it hard to see
the relationships between them. We can store the alias information in
the YAML file and automatically map the aliases to the actual labels
when loading the file.