Apply CPUQuota and cpu.max.burst to slice
When burst-related parameters are set in prep.json, they are applied
to the slice for a VM service.

As-is, this is intended have no effect, as all VMs sold today are
dedicated-core.  Furthermore, these slice settings are compatible and
a no-op with Ubuntu 22.04 as a degenerate case, as it does not have a
new enough kernel to support `cpu.max.burst`.