Create VmHostCpu records when setting up a VmHost.
When host is created we create a record for each of the host's cpus. A
number of cpus is set aside for SPDK, and the rest are marked as
available. Each VmHostSlide can get a set of cpus assigned to it and
effectively reserved for that slice. That functionality is not fully
used yet but will be expanded upon in subsequent commits.