Fix error if user visits verify-account-resend page directly
Usually, this page is displayed on login and account creation if the
account is unverified.  In that case, we know the account and can
determine whether the verification email was recently sent.  If the
page is accessed directly, we may not know if the verification email
was sent recently or not.  Previously, this raised a NoMethodError
internally on Rodauth.

I've made changes to Rodauth to not raise an exception in this case,
but I don't think the behavior is what we want.  If a user visits
the page while not logged in, when we cannot determine whether an
email was recently sent, we should show the warning.  So change
the behavior, and make the following changes:

* Change displayed text from 300 seconds to 5 minutes to make it
  easier on the user

* Move verify_account_email_recently_sent? check to the template,
  and store the value to avoid redundant queries

* Fix invalid HTML (nested `<p>` tags)

After changes:

Line Coverage: 99.87% (11708 / 11723)
Branch Coverage: 98.51% (3099 / 3146)