Unify empty states in tables
Some pages display a message in the table body when empty, while others
show a full-page empty state. This PR unifies the empty states to
consistently show a message in the table body. It removes the
conditional wrapping the table. The new method of displaying the empty
state is also more consistent with the rest of the app and includes the
page header.