Remove unneeded branches in two_factor_auth page
This page is only shown if both otp and webauthn are setup for
the account.  If otp is setup and webauthn is not, request
is redirected to otp page.  If webauthn is setup and otp is
not, request is redirected to webauthn page.

This should be reverted if we add an multifactor option
(beyond recovery codes, which are deliberately not shown
as an option on this page).

After changes:

Line Coverage: 100.0% (11677 / 11677)
Branch Coverage: 99.65% (3096 / 3107)