Keep submitted checkboxes when attempting to remove all members from Admin subject tag
Instead of a normal redirect, raise a Sequel::ValidationFailed exception,
which the error handler will handle, and will result in `flash["old"]`
being set.

It wasn't working even after that fix, because remove_checked
is an array of UUIDs, not an array of UBIDs, so convert
from UBID to UUID when calling `remove_checked&.include?`.

After changes:

Line Coverage: 100.0% (11673 / 11673)
Branch Coverage: 99.71% (3096 / 3105)