Switch from multi_insert to import in DnsZone#delete_record
Dataset#multi_insert just converts the array of hashes to an
array of arrays and runs Dataset#import, so this is just a more
efficient approach, no behavior change.  It's nice how the
columns line up, though.

If we had a database function for generating a uuid which was a
valid ubid for the object type, we could change the
records.select_map to a records.select, so it could use a single
query:

INSERT INTO ... SELECT ...

instead of 2 queries:

SELECT ...
INSERT INTO ... VALUES ...