Add project_id to tables that should only be associated with a single project
Previously, the following tables could be associated to 0 or more
projects:

* api_key
* firewall
* load_balancer
* minio_cluster
* private_subnet
* vm

It only makes sense for these to be associated to a single project,
so this adds a project_id column to each.  This is the first step
of the conversion process, which will eventually see the access_tag
table only be used for a project <-> accounts join table.