Improve validation logic for setting up DNS Server VM
This commit improves the validation logic of DNS Server VM setup.

1. Validate existing VMs in .assemble
In case of any sync issues within the VMs of a DNS server, the prog
fails earlier in the .assemble step to make investigation easier. We
hit such a case, although it should be pretty rare.

2. Validation should ignore the serial number of SOA records as the
number is incremented with every update and it's OK to see different
values on different VMs.