Fail closed if ubid without related class is given as object during authorization query
Previously, if a ubid without a related class was passed,
UBID.class_for_ubid would return nil, and it would not check that
the object was related to the project.  This potentially affected
ApiKey, which uses the TYPE_ETC type.  Treat 'et' UBIDs as being
ApiKey, since nothing else used as objects in authorization queries
uses that type.

Additionally, fail closed by making the matched policies dataset
always return no rows if the UBID is not related to the class,
which will result in the authorization being denied.