Introduce Kubernetes Nodepools
This change adds the concept of Kubernetes Nodepools.

Each kubernetes nodepool is a set of VMs that act as the nodes that run
the actual application workloads.

A cluster have multiple nodepools with different size and
characteristics. The logic in this PR is also implemented this way.
However, product-wise, we may hide this feature for now and default to a
single nodepool.

A kubernetes cluster without a nodepool is kinda useless at this point.

Nodepools should be created separately from the cluster. However, a
nodepool nexus strand would wait until its cluster is stable.

Co-authored-by: Eren Başak <eren@ubicloud.com>
Co-authored-by: mohi-kalantari <mohi.kalantari1@gmail.com>