Set docker mirror client code for the runners on the test host
In order to test docker registry mirror performance, registry
mirror server is set up on one of the host. It is attached to a
load balancer to have dns address and certificates automatically. It
may also help in the future if we will want to load balance image
pull requests.

Updating docker and buildkit config here to use that server for the
runners provisioned on the test host, so we can assess initial reliability
and performance for runners on that host.