Ensure all boot images are tested
Previously, we randomly selected a boot image for VM tests. This means
that if a specific boot image broke, we might miss it since we don’t
test all boot images all the time.

However, the green mark can mislead developers into thinking everything
is fine.

We should test all boot images to ensure they are functioning properly.
This PR creates virtual machines to test at least a minimum number of
boot images.

If the number of provided boot images is less than 3, it will repeat the
boot images to reach a total of 3. This way, we will create at least 3
virtual machines for testing.