Drop columns from access_tag, and drop applied_tag and access_policy
This makes access_tag a pure project <-> accounts join table
(modulo the created_at column to see when the two were
associated).

This removes non-account entries from the access_tag table,
because that is required for the foreign key constraint to work.

While here, drop the applied_tag and access_policy tables, as they
are no longer used.