Create SA, Secret andClusterRolebinding for kubeconfig generation
We will avoid handing over the admin kubeconfig, instead we will create
Secrets, ClusterRolebinding and ServiceAccount in order to generate JWT
tokens.

In the UI, whenever a kubeconfig retrieval is requested, we will extract
the token from the secret and then assemble the kubeconfig

The script will run once after the kubeadm init.