Adding VmHostSlideNexus prog
Adding a new prog for VmHostSlice. A slice controls the allocation of CPUs and memory to one or more VMs. When created and associated with a VmHost, it creates a SystemCtl slice object (and underlying cgroup) on the host. It takes as an argument a set of CPUs to be affinitized to the group and an amount of memory to keep track of.

Also adding support for availability monitoring, following the standard pattern for ./bin/monitor script.

The VmHostSlice checks for three conditions:
- if the slice systemd unit is available
- if the allowed cpus is set to what is configured
- if the slice partition type is 'root' which indicates exclusive cpuset